+++ apps/api/v1/connection/google_cloud/views.py
@@ -15,7 +15,7 @@ from apps.api.v1.utils.api_permissions import MemberPermissions
 from apps.console.node.models import CoreOVHCA, CoreGoogleCloud, CoreNode
 from .filters import CoreGoogleCloudFilter
 from .serializers import CoreGoogleCloudConnectionReadSerializer, CoreGoogleCloudConnectionWriteSerializer
-from ..._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
+from apps._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
     IntegrationValidationError
 from ...utils.api_filters import DateRangeFilter
 from ...utils.api_serializers import ReadWriteSerializerMixin
