+++ apps/api/v1/connection/hetzner/views.py
@@ -13,7 +13,7 @@ from apps.console.node.models import CoreHetzner, CoreNode
 from .filters import CoreHetznerFilter
 from .permissions import CoreHetznerViewPermissions
 from .serializers import CoreHetznerConnectionReadSerializer, CoreHetznerConnectionWriteSerializer
-from ..._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
+from apps._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
     IntegrationValidationError
 from ...utils.api_filters import DateRangeFilter
 from ...utils.api_serializers import ReadWriteSerializerMixin
