+++ apps/api/v1/connection/ovh_eu/views.py
@@ -15,7 +15,7 @@ from apps.api.v1.utils.api_permissions import MemberPermissions
 from apps.console.node.models import CoreOVHEU, CoreNode
 from .filters import CoreOVHEUFilter
 from .serializers import CoreOVHEUConnectionReadSerializer, CoreOVHEUConnectionWriteSerializer
-from ..._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
+from apps._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
     IntegrationValidationError
 from ...utils.api_filters import DateRangeFilter
 from ...utils.api_serializers import ReadWriteSerializerMixin
