+++ apps/api/v1/connection/upcloud/views.py
@@ -13,7 +13,7 @@ from apps.console.node.models import CoreUpCloud, CoreNode
 from .filters import CoreUpCloudFilter
 from .permissions import CoreUpCloudViewPermissions
 from .serializers import CoreUpCloudConnectionReadSerializer, CoreUpCloudConnectionWriteSerializer
-from ..._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
+from apps._tasks.exceptions import NodeConnectionErrorEligibleObjects, IntegrationValidationFailed, \
     IntegrationValidationError
 from ...utils.api_filters import DateRangeFilter
 from ...utils.api_serializers import ReadWriteSerializerMixin
