+++ apps/api/v1/node/views.py
@@ -12,16 +12,16 @@ from apps.api.v1.utils.api_permissions import MemberPermissions
 from apps.console.node.models import CoreNode
 from .filters import CoreNodeFilter
 from .serializers import CoreNodeSerializer
-from .._tasks.exceptions import (
+from apps._tasks.exceptions import (
     SnapshotCreateMissingParams,
     SnapshotCreateError,
     SnapshotCreateNodeValidationFailed,
     SnapshotCreateNodeNotActive, NodeValidationFailed, AccountNotGoodStanding,
 )
-from .._tasks.integration.basecamp import backup_basecamp
-from .._tasks.integration.website import backup_website
+from apps._tasks.integration.basecamp import backup_basecamp
+from apps._tasks.integration.website import backup_website
 from ..utils.api_filters import DateRangeFilter
-from apps.api.v1._tasks.helper.tasks import node_delete_requested
+from apps._tasks.helper.tasks import node_delete_requested
 from ..utils.api_helpers import delete_snar_file
