+++ apps/api/v1/storage/scaleway/views.py
@@ -18,7 +18,7 @@ from apps.console.storage.models import CoreStorage
 from .filters import CoreStorageScalewayFilter
 from .permissions import CoreStorageScalewayPermissions
 from .serializers import CoreStorageReadSerializer, CoreStorageWriteSerializer
-from ..._tasks.exceptions import StorageValidationFailed
+from apps._tasks.exceptions import StorageValidationFailed
 from ...utils.api_filters import DateRangeFilter
 from ...utils.api_helpers import get_start_end_of_previous_day
 from ...utils.api_serializers import ReadWriteSerializerMixin
